[API]: Move soundcoe.cpp implementation into the header
